+++ src/color.rs
@@ -8,7 +8,7 @@ pub fn color(ray: &Ray, world: &HittableList, depth: i32) -> Vec3 {
         return Vec3::new(0.0, 0.0, 0.0);
     }
 
-    match world.hit(ray, 0.001, f64::MAX) {
+    match world.hit(ray, 0.001, f32::MAX) {
         Some(hit) => {
             if let Some((attenuation, scattered)) = hit.material.scatter(ray, &hit) {
                 attenuation * color(&scattered, world, depth - 1)
