+++ src/sphere.rs
@@ -5,18 +5,18 @@ use crate::vec3::Vec3;
 
 pub struct Sphere {
     pub center: Vec3,
-    pub radius: f64,
+    pub radius: f32,
     pub material: Box<dyn Material>,
 }
 
 impl Sphere {
-    pub fn new(center: Vec3, radius: f64, material: Box<dyn Material>) -> Sphere {
+    pub fn new(center: Vec3, radius: f32, material: Box<dyn Material>) -> Sphere {
         Sphere { center, radius, material }
     }
 }
 
 impl Hittable for Sphere {
-    fn hit(&self, ray: &Ray, t_min: f64, t_max: f64) -> Option<HitRecord<'_>> {
+    fn hit(&self, ray: &Ray, t_min: f32, t_max: f32) -> Option<HitRecord<'_>> {
         let oc = ray.origin() - self.center;
         let a = ray.direction().dot(ray.direction());
         let b = oc.dot(ray.direction());
